Move system manager to water resource manager
