--- src/API/WesternStatesWater.WestDaat.Contracts.Client/ISystemManager.cs
@@ -1,12 +1,0 @@
-﻿using GeoJSON.Text.Feature;
-using WesternStatesWater.WestDaat.Common;
-
-namespace WesternStatesWater.WestDaat.Contracts.Client
-{
-    public interface ISystemManager : IServiceContractBase
-    {
-        Task<DashboardFilters> LoadFilters();
-
-        FeatureCollection GetRiverBasinPolygonsByName(string[] basinNames);
-    }
-}
