+++ src/API/WesternStatesWater.WestDaat.Contracts.Client/IWaterResourceManager.cs
@@ -2,57 +2,60 @@ using GeoJSON.Text.Feature;
 using System.IO;
 using WesternStatesWater.WestDaat.Common;
 
-namespace WesternStatesWater.WestDaat.Contracts.Client
+namespace WesternStatesWater.WestDaat.Contracts.Client;
+
+/// <summary>
+/// Defines operations for searching, viewing, and analyzing water rights, sites, allocations, 
+/// and overlay data. Implementations of this interface are intended to support comprehensive 
+/// data queries and provide insights into water resource management.
+/// </summary>
+public interface IWaterResourceManager : IServiceContractBase
 {
-    /// <summary>
-    /// Defines operations for searching, viewing, and analyzing water rights, sites, allocations, 
-    /// and overlay data. Implementations of this interface are intended to support comprehensive 
-    /// data queries and provide insights into water resource management.
-    /// </summary>
-    public interface IWaterResourceManager : IServiceContractBase
-    {
-        Task<AnalyticsSummaryInformationResponse> GetAnalyticsSummaryInformation(WaterRightsSearchCriteria searchRequest);
+    Task<WaterRightsSearchResults> FindWaterRights(WaterRightsSearchCriteriaWithPaging searchRequest);
 
-        Task<FeatureCollection> GetWaterRightsEnvelope(WaterRightsSearchCriteria searchRequest);
+    Task<AnalyticsSummaryInformationResponse> GetAnalyticsSummaryInformation(WaterRightsSearchCriteria searchRequest);
 
-        Task<WaterRightsSearchResults> FindWaterRights(WaterRightsSearchCriteriaWithPaging searchRequest);
+    Task<FeatureCollection> GetNldiFeatures(double latitude, double longitude, NldiDirections directions, NldiDataPoints dataPoints);
 
-        Task<FeatureCollection> GetNldiFeatures(double latitude, double longitude, NldiDirections directions, NldiDataPoints dataPoints);
+    Task<OverlayDetails> GetOverlayDetails(string overlayUuid);
 
-        Task<WaterRightDetails> GetWaterRightDetails(string allocationUuid);
+    Task<List<OverlayTableEntry>> GetOverlayInfoById(string reportingUnitUuid);
 
-        Task<List<SiteInfoListItem>> GetWaterRightSiteInfoList(string allocationUuid);
+    FeatureCollection GetRiverBasinPolygonsByName(string[] basinNames);
 
-        Task<List<WaterSourceInfoListItem>> GetWaterRightSourceInfoList(string allocationUuid);
+    Task<SiteDetails> GetSiteDetails(string siteUuid);
 
-        Task<FeatureCollection> GetWaterRightSiteLocations(string allocationUuid);
+    Task<SiteDigest> GetSiteDigest(string siteUuid);
 
-        Task<string> GetWaterAllocationSiteGeoconnexIntegrationData(string siteUuid);
+    Task<List<MethodInfoListItem>> GetSiteMethodInfoListByUuid(string siteUuid);
 
-        Task<SiteDetails> GetSiteDetails(string siteUuid);
+    Task<SiteUsage> GetSiteUsageBySiteUuid(string siteUuid);
 
-        Task<List<WaterRightsDigest>> GetWaterRightsDigestsBySite(string siteUuid);
+    Task<List<VariableInfoListItem>> GetSiteVariableInfoListByUuid(string siteUuid);
 
-        Task<SiteDigest> GetSiteDigest(string siteUuid);
+    Task<string> GetWaterAllocationSiteGeoconnexIntegrationData(string siteUuid);
 
-        Task<Feature> GetWaterSiteLocation(string siteUuid);
+    Task<WaterRightDetails> GetWaterRightDetails(string allocationUuid);
 
-        Task<List<WaterSourceInfoListItem>> GetWaterSiteSourceInfoListByUuid(string siteUuid);
+    Task<List<WaterRightsDigest>> GetWaterRightsDigestsBySite(string siteUuid);
 
-        Task<List<WaterRightInfoListItem>> GetWaterSiteRightsInfoListByUuid(string siteUuid);
+    Task<FeatureCollection> GetWaterRightsEnvelope(WaterRightsSearchCriteria searchRequest);
 
-        Task WaterRightsAsZip(Stream responseStream, WaterRightsSearchCriteriaWithFilterUrl searchRequest);
+    Task<List<WaterRightInfoListItem>> GetWaterRightsInfoListByReportingUnitUuid(string reportingUnitUuid);
 
-        Task<SiteUsage> GetSiteUsageBySiteUuid(string siteUuid);
+    Task<List<SiteInfoListItem>> GetWaterRightSiteInfoList(string allocationUuid);
 
-        Task<List<VariableInfoListItem>> GetSiteVariableInfoListByUuid(string siteUuid);
-        
-        Task<OverlayDetails> GetOverlayDetails(string overlayUuid);
-        
-        Task<List<OverlayTableEntry>> GetOverlayInfoById(string reportingUnitUuid);
-        
-        Task<List<MethodInfoListItem>> GetSiteMethodInfoListByUuid(string siteUuid);
-        
-        Task<List<WaterRightInfoListItem>> GetWaterRightsInfoListByReportingUnitUuid(string reportingUnitUuid);
-    }
+    Task<FeatureCollection> GetWaterRightSiteLocations(string allocationUuid);
+
+    Task<List<WaterSourceInfoListItem>> GetWaterRightSourceInfoList(string allocationUuid);
+
+    Task<Feature> GetWaterSiteLocation(string siteUuid);
+
+    Task<List<WaterRightInfoListItem>> GetWaterSiteRightsInfoListByUuid(string siteUuid);
+
+    Task<List<WaterSourceInfoListItem>> GetWaterSiteSourceInfoListByUuid(string siteUuid);
+
+    Task<DashboardFilters> LoadFilters();
+
+    Task WaterRightsAsZip(Stream responseStream, WaterRightsSearchCriteriaWithFilterUrl searchRequest);
 }
